+++ index.d.ts
@@ -71,8 +71,12 @@ export class Resvg {
   /**
    * Use a given `BBox` to crop the svg. Currently this method simply changes
    * the viewbox/size of the svg and do not move the elements for simplicity
+   * # Arguments
+   * `bbox` - The bounding box to crop to
+   * `padding` - Optional bleed area around the crop box (default: 0.0)
    */
-  cropByBBox(bbox: BBox): void
+  // cropByBBox(bbox: BBox): void
+  cropByBBox(bbox: BBox, padding?: number): void
 
   imagesToResolve(): Array<string>
   resolveImage(href: string, buffer: Buffer): void